Refractor fund views
